refac & init tests
