+++ src/test/java/process/TestSchedulesUpdateProcess.java
@@ -1,0 +1,15 @@
+package process;
+
+/**
+ * Classe de test de {@link SchedulesUpdateProcess}.
+ */
+public class TestSchedulesUpdateProcess {
+
+  /*
+  Séparer :
+    - la récupération des données : tester avec un lien valide en variable d'environnement
+    - le mapping des données : écrire des données ical de test
+    - les traitements : créer modèles de test
+    - la purge des sessions mises à jour : créer modèles de test
+   */
+}
